Rename data processing modeule to pipeline
